Redo memory_array_dual constraint assignment
The previously tricksy code using custom attributes and TCL to walk
backwards to find any RAMD instances is much too brittle.  Change this
instead to simply use a custom name format, specifically we now search for
    NAME =~ */memory_array_dual_memory* && REF_NAME =~ RAMD*
This seems to work properly.

This will fail if the FIFO is assigned to registers, and we have now
broken the special case for in_fifo.